rename: entities/interview-question 구조 전환
