+++ client/src/entities/interview_question/api/queries.ts
@@ -1,5 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { QuestionType } from '@/src/entities/interview_question/type';
+
+import { QuestionType } from '@/src/entities/interview_question/model/type';
+
 import interview from './queryKeys';
 import { getInterViewQuestionList } from './apis';
