+++ client/src/entities/interview_question/api/queryKeys.ts
@@ -1,4 +1,4 @@
-import { QuestionType } from '@/src/entities/interview_question/type';
+import { QuestionType } from '@/src/entities/interview_question/model/type';
 
 const interview = {
   all: ['interview'],
