+++ client/src/entities/interview_question/model/type.ts
